Add files via upload
Latest version of files with action code. The action is throwing an error when I run the rasa server 'AttributeError: 'str' object has no attribute 'name' "
I have added a new entity restaurant and action for the ActionNeo4j actions class just to test this with a separate intent to run the action class.

The NLU has the same issues in terms of of not returning the description query attribute from equipment for example.